requirements.txt
markdown==3.5.2
python-frontmatter==1.0.0
jinja2==3.1.3
watchdog==3.0.0 